[iPad] Fixes loading wrong game after quitting game in another window
Previously quitting a scene session didn’t quit emulation, which meant the EmulatorBridge state was never reset. If the user then loaded a new game for the same system, it could potentially just resume the old game instead (and corrupt save data in the process).

We now explicitly quit emulation (and make an auto-save) when quitting a scene session.